refactor: move code to internal packages
