+++ internal/ui/nineslice.go
@@ -1,7 +1,7 @@
 //go:build !headless
 // +build !headless
 
-package main
+package ui
 
 import (
 	"image"
@@ -13,8 +13,7 @@ import (
 )
 
 // NineSlice represents an image that can be drawn using the nine-slice
-// technique. The Corner field specifies the size of each corner region
-// in pixels.
+// technique. The Corner field specifies the size of each corner region in pixels.
 type NineSlice struct {
 	Image  *ebiten.Image
 	Corner int
@@ -27,7 +26,6 @@ func LoadNineSlice(path string, corner int) (*NineSlice, error) {
 	img, _, err := ebitenutil.NewImageFromFile(path)
 	if err != nil {
 		log.Printf("nine-slice load error: %v", err)
-		// create a visible placeholder so the game can continue running
 		ph := ebiten.NewImage(corner*2, corner*2)
 		ph.Fill(color.RGBA{255, 0, 255, 255})
 		return &NineSlice{Image: ph, Corner: corner}, err
@@ -45,7 +43,6 @@ func (ns *NineSlice) Draw(dst *ebiten.Image, rect image.Rectangle) {
 	cw := ns.Corner
 	sw, sh := ns.Image.Size()
 	if rect.Dx() < 2*cw || rect.Dy() < 2*cw {
-		// rectangle too small, just draw the whole image scaled
 		op := &ebiten.DrawImageOptions{}
 		op.GeoM.Scale(float64(rect.Dx())/float64(sw), float64(rect.Dy())/float64(sh))
 		op.GeoM.Translate(float64(rect.Min.X), float64(rect.Min.Y))
@@ -53,7 +50,6 @@ func (ns *NineSlice) Draw(dst *ebiten.Image, rect image.Rectangle) {
 		return
 	}
 
-	// helper to draw a subimage with scaling
 	drawPart := func(src image.Rectangle, x, y, w, h int) {
 		sub := ns.Image.SubImage(src).(*ebiten.Image)
 		op := &ebiten.DrawImageOptions{}
@@ -64,20 +60,17 @@ func (ns *NineSlice) Draw(dst *ebiten.Image, rect image.Rectangle) {
 
 	left, top := rect.Min.X, rect.Min.Y
 	right, bottom := rect.Max.X, rect.Max.Y
-	// corners
 	drawPart(image.Rect(0, 0, cw, cw), left, top, cw, cw)
 	drawPart(image.Rect(sw-cw, 0, sw, cw), right-cw, top, cw, cw)
 	drawPart(image.Rect(0, sh-cw, cw, sh), left, bottom-cw, cw, cw)
 	drawPart(image.Rect(sw-cw, sh-cw, sw, sh), right-cw, bottom-cw, cw, cw)
 
-	// edges
 	w := rect.Dx() - 2*cw
 	h := rect.Dy() - 2*cw
-	drawPart(image.Rect(cw, 0, sw-cw, cw), left+cw, top, w, cw)           // top
-	drawPart(image.Rect(cw, sh-cw, sw-cw, sh), left+cw, bottom-cw, w, cw) // bottom
-	drawPart(image.Rect(0, cw, cw, sh-cw), left, top+cw, cw, h)           // left
-	drawPart(image.Rect(sw-cw, cw, sw, sh-cw), right-cw, top+cw, cw, h)   // right
+	drawPart(image.Rect(cw, 0, sw-cw, cw), left+cw, top, w, cw)
+	drawPart(image.Rect(cw, sh-cw, sw-cw, sh), left+cw, bottom-cw, w, cw)
+	drawPart(image.Rect(0, cw, cw, sh-cw), left, top+cw, cw, h)
+	drawPart(image.Rect(sw-cw, cw, sw, sh-cw), right-cw, top+cw, cw, h)
 
-	// center
 	drawPart(image.Rect(cw, cw, sw-cw, sh-cw), left+cw, top+cw, w, h)
 }
